+++ supreme/common/src/androidJvmMain/kotlin/JvmExtensions.kt
@@ -3,6 +3,7 @@ package at.asitplus.attestation.android
 import at.asitplus.catchingUnwrapped
 import at.asitplus.signum.indispensable.toKmpCertificate
 
+
 /**
  * Tries to parse an [AttestationKeyDescription] certificate extension, if present.
  * Never throws.
@@ -10,6 +11,15 @@ import at.asitplus.signum.indispensable.toKmpCertificate
 val java.security.cert.X509Certificate.androidAttestationExtension: AttestationKeyDescription?
     get() = catchingUnwrapped {
         toKmpCertificate().getOrNull()?.androidAttestationExtension
+    }.getOrElse {
+        null
+    }
+
+val List<java.security.cert.X509Certificate>.androidAttestationExtension: AttestationKeyDescription?
+    get() = catchingUnwrapped {
+        mapNotNull { it.toKmpCertificate().getOrNull() }.let {
+            if (it.size != this.size) null else it.androidAttestationExtension
+        }
     }.getOrElse {
         null
     }